Rename submission_schema -> nmdc_submission_schema
