+++ .github/workflows/main.yaml
@@ -2,7 +2,7 @@
 # https://docs.github.com/en/actions/guides/building-and-testing-python
 # https://github.com/snok/install-poetry#workflows-and-tips
 
-name: Build and test submission_schema
+name: Build and test nmdc_submission_schema
 
 on: [pull_request]
