+++ README.md
@@ -2,6 +2,8 @@
 
 The home of the NMDC submission schema. *Not* the home of sheets_and_friends. *Not* a GH pages host of NMDC DataHarmonizer interfaces.
 
+Note that the while this repo is named `submission-schema`, the generated artifacts are named `nmdc_submission_schema` for disambiguation purposes when publishing to PyPI.
+
 ## Website
 
 * [https://microbiomedata.github.io/submission-schema](https://microbiomedata.github.io/submission-schema)
@@ -11,9 +13,9 @@ The home of the NMDC submission schema. *Not* the home of sheets_and_friends. *N
 * [examples/](examples/) - example data
 * [project/](project/) - project files (do not edit these)
 * [src/](src/) - source files (edit these)
-    * [submission_schema](src/submission_schema)
-        * [schema](src/submission_schema/schema) -- LinkML schema (edit this)
-* [datamodel](src/submission_schema/datamodel) -- Generated python datamodel
+    * [nmdc_submission_schema](src/nmdc_submission_schema)
+        * [schema](src/nmdc_submission_schema/schema) -- LinkML schema (edit this)
+* [datamodel](src/nmdc_submission_schema/datamodel) -- Generated python datamodel
 * [tests](tests/) - python tests
 
 ## Developer Documentation
