+++ examples/README.md
@@ -1,5 +1,5 @@
-# Examples of use of submission_schema
+# Examples of use of nmdc_submission_schema
 
-This folder contains example data conforming to submission_schema
+This folder contains example data conforming to nmdc_submission_schema
 
 The source for these is in [src/data](../src/data/examples)
