+++ src/nmdc_submission_schema/datamodel/__init__.py
@@ -1,0 +1,1 @@
+from .nmdc_submission_schema import *
