--- src/submission_schema/datamodel/__init__.py
@@ -1,1 +1,0 @@
-from .submission_schema import *
